Use a fork of BottomNavigator able to properly handle having the bottom navigation only in MainFragment
