+++ simplified-main/src/main/java/org/nypl/simplified/main/MainFragmentNavigationDelegate.kt
@@ -1,28 +1,15 @@
 package org.nypl.simplified.main
 
-import android.content.Context
-import android.content.res.ColorStateList
-import android.graphics.Color
-import androidx.annotation.IdRes
 import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.google.android.material.bottomnavigation.LabelVisibilityMode
-import com.io7m.junreachable.UnreachableCodeException
-import com.pandora.bottomnavigator.BottomNavigator
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
 import io.reactivex.disposables.CompositeDisposable
-import org.joda.time.DateTime
-import org.nypl.simplified.accounts.api.AccountProviderType
-import org.nypl.simplified.accounts.database.api.AccountType
-import org.nypl.simplified.accounts.registry.api.AccountProviderRegistryType
 import org.nypl.simplified.books.api.Book
 import org.nypl.simplified.books.api.BookFormat
 import org.nypl.simplified.books.api.BookID
 import org.nypl.simplified.buildconfig.api.BuildConfigurationServiceType
-import org.nypl.simplified.feeds.api.FeedBooksSelection
 import org.nypl.simplified.feeds.api.FeedEntry
-import org.nypl.simplified.feeds.api.FeedFacet
 import org.nypl.simplified.profiles.controller.api.ProfilesControllerType
 import org.nypl.simplified.ui.accounts.AccountFragment
 import org.nypl.simplified.ui.accounts.AccountFragmentParameters
@@ -32,7 +19,6 @@ import org.nypl.simplified.ui.accounts.AccountListRegistryFragment
 import org.nypl.simplified.ui.accounts.saml20.AccountSAML20Fragment
 import org.nypl.simplified.ui.accounts.saml20.AccountSAML20FragmentParameters
 import org.nypl.simplified.ui.catalog.CatalogFeedArguments
-import org.nypl.simplified.ui.catalog.CatalogFeedOwnership
 import org.nypl.simplified.ui.catalog.CatalogFragmentBookDetail
 import org.nypl.simplified.ui.catalog.CatalogFragmentBookDetailParameters
 import org.nypl.simplified.ui.catalog.CatalogFragmentFeed
@@ -43,262 +29,21 @@ import org.nypl.simplified.ui.errorpage.ErrorPageParameters
 import org.nypl.simplified.ui.navigation.tabs.R
 import org.nypl.simplified.ui.navigation.tabs.TabbedNavigationCommand
 import org.nypl.simplified.ui.navigation.tabs.TabbedNavigationController
-import org.nypl.simplified.ui.profiles.ProfileTabFragment
+import org.nypl.simplified.ui.navigation.tabs.TabbedNavigator
 import org.nypl.simplified.ui.settings.SettingsFragmentCustomOPDS
 import org.nypl.simplified.ui.settings.SettingsFragmentDebug
-import org.nypl.simplified.ui.settings.SettingsFragmentMain
-import org.nypl.simplified.ui.theme.ThemeControl
 import org.nypl.simplified.viewer.api.Viewers
 import org.nypl.simplified.viewer.spi.ViewerPreferences
 import org.slf4j.LoggerFactory
 import java.net.URI
 
-/**
- * A tabbed navigation system based on Pandora's BottomNavigator.
- *
- * @see BottomNavigator
- * @see BottomNavigationView
- */
-
-
-class MainFragmentNavigationDelegate(
+internal class MainFragmentNavigationDelegate(
   private val activity: AppCompatActivity,
   private val navigationController: TabbedNavigationController,
-  private val navigator: BottomNavigator,
-  private val settingsConfiguration: BuildConfigurationServiceType,
+  private val navigator: TabbedNavigator,
   private val profilesController: ProfilesControllerType,
-) {
-
-  companion object {
-
-    private val logger = LoggerFactory.getLogger(MainFragmentNavigationDelegate::class.java)
-
-    /**
-     * Create a new tabbed navigation controller. The controller will load fragments into the
-     * fragment container specified by [fragmentContainerId], using the Pandora BottomNavigator
-     * view [navigationView].
-     */
-
-    fun create(
-      activity: AppCompatActivity,
-      navigationController: TabbedNavigationController,
-      accountProviders: AccountProviderRegistryType,
-      profilesController: ProfilesControllerType,
-      settingsConfiguration: BuildConfigurationServiceType,
-      @IdRes fragmentContainerId: Int,
-      navigationView: BottomNavigationView,
-    ): MainFragmentNavigationDelegate {
-      logger.debug("creating bottom navigator")
-      val navigator =
-        BottomNavigator.onCreate(
-          fragmentContainer = fragmentContainerId,
-          bottomNavigationView = navigationView,
-          rootFragmentsFactory = mapOf(
-            R.id.tabCatalog to {
-              createCatalogFragment(
-                context = activity,
-                id = R.id.tabCatalog,
-                feedArguments = catalogFeedArguments(
-                  activity,
-                  profilesController,
-                  accountProviders.defaultProvider
-                )
-              )
-            },
-            R.id.tabBooks to {
-              createBooksFragment(
-                context = activity,
-                id = R.id.tabBooks,
-                profilesController = profilesController,
-                settingsConfiguration = settingsConfiguration,
-                defaultProvider = accountProviders.defaultProvider
-              )
-            },
-            R.id.tabHolds to {
-              createHoldsFragment(
-                context = activity,
-                id = R.id.tabHolds,
-                profilesController = profilesController,
-                settingsConfiguration = settingsConfiguration,
-                defaultProvider = accountProviders.defaultProvider
-              )
-            },
-            R.id.tabSettings to {
-              createSettingsFragment(R.id.tabSettings)
-            },
-            R.id.tabProfile to {
-              ProfileTabFragment()
-            }
-          ),
-          defaultTab = R.id.tabCatalog,
-          activity = activity
-        )
-
-      navigationView.itemIconTintList = colorStateListForTabs(activity)
-      navigationView.itemTextColor = colorStateListForTabs(activity)
-      navigationView.labelVisibilityMode = LabelVisibilityMode.LABEL_VISIBILITY_LABELED
-
-      return MainFragmentNavigationDelegate(
-        activity = activity,
-        navigationController = navigationController,
-        navigator = navigator,
-        settingsConfiguration = settingsConfiguration,
-        profilesController = profilesController,
-      )
-    }
-
-    private fun currentAge(
-      profilesController: ProfilesControllerType
-    ): Int {
-      return try {
-        val profile = profilesController.profileCurrent()
-        profile.preferences().dateOfBirth?.yearsOld(DateTime.now()) ?: 1
-      } catch (e: Exception) {
-        logger.error("could not retrieve profile age: ", e)
-        1
-      }
-    }
-
-    private fun pickDefaultAccount(
-      profilesController: ProfilesControllerType,
-      defaultProvider: AccountProviderType
-    ): AccountType {
-      val profile = profilesController.profileCurrent()
-      val mostRecentId = profile.preferences().mostRecentAccount
-      if (mostRecentId != null) {
-        try {
-          return profile.account(mostRecentId)
-        } catch (e: Exception) {
-          logger.error("stale account: ", e)
-        }
-      }
-
-      val accounts = profile.accounts().values
-      return when {
-        accounts.size > 1 -> {
-          // Return the first account created from a non-default provider
-          accounts.first { it.provider.id != defaultProvider.id }
-        }
-        accounts.size == 1 -> {
-          // Return the first account
-          accounts.first()
-        }
-        else -> {
-          // There should always be at least one account
-          throw UnreachableCodeException()
-        }
-      }
-    }
-
-    private fun catalogFeedArguments(
-      context: Context,
-      profilesController: ProfilesControllerType,
-      defaultProvider: AccountProviderType
-    ): CatalogFeedArguments.CatalogFeedArgumentsRemote {
-      val age = currentAge(profilesController)
-      val account = pickDefaultAccount(profilesController, defaultProvider)
-      return CatalogFeedArguments.CatalogFeedArgumentsRemote(
-        ownership = CatalogFeedOwnership.OwnedByAccount(account.id),
-        feedURI = account.catalogURIForAge(age),
-        isSearchResults = false,
-        title = context.getString(R.string.tabCatalog)
-      )
-    }
-
-    private fun colorStateListForTabs(context: FragmentActivity): ColorStateList {
-      val states =
-        arrayOf(
-          intArrayOf(android.R.attr.state_checked),
-          intArrayOf(-android.R.attr.state_checked)
-        )
-
-      val colors =
-        intArrayOf(
-          ThemeControl.resolveColorAttribute(context.theme, R.attr.colorPrimary),
-          Color.DKGRAY
-        )
-
-      return ColorStateList(states, colors)
-    }
-
-    private fun createSettingsFragment(id: Int): Fragment {
-      logger.debug("[{}]: creating settings fragment", id)
-      return SettingsFragmentMain()
-    }
-
-    private fun createHoldsFragment(
-      context: Context,
-      id: Int,
-      profilesController: ProfilesControllerType,
-      settingsConfiguration: BuildConfigurationServiceType,
-      defaultProvider: AccountProviderType
-    ): Fragment {
-      logger.debug("[{}]: creating holds fragment", id)
-
-      /*
-       * SIMPLY-2923: Filter by the default account until 'All' view is approved by UX.
-       */
-
-      val filterAccountId =
-        if (settingsConfiguration.showBooksFromAllAccounts) {
-          null
-        } else {
-          pickDefaultAccount(profilesController, defaultProvider).id
-        }
-
-      return CatalogFragmentFeed.create(
-        CatalogFeedArguments.CatalogFeedArgumentsLocalBooks(
-          filterAccount = filterAccountId,
-          ownership = CatalogFeedOwnership.CollectedFromAccounts,
-          searchTerms = null,
-          selection = FeedBooksSelection.BOOKS_FEED_HOLDS,
-          sortBy = FeedFacet.FeedFacetPseudo.Sorting.SortBy.SORT_BY_TITLE,
-          title = context.getString(R.string.tabHolds)
-        )
-      )
-    }
-
-    private fun createBooksFragment(
-      context: Context,
-      id: Int,
-      profilesController: ProfilesControllerType,
-      settingsConfiguration: BuildConfigurationServiceType,
-      defaultProvider: AccountProviderType
-    ): Fragment {
-      logger.debug("[{}]: creating books fragment", id)
-
-      /*
-       * SIMPLY-2923: Filter by the default account until 'All' view is approved by UX.
-       */
-
-      val filterAccountId =
-        if (settingsConfiguration.showBooksFromAllAccounts) {
-          null
-        } else {
-          pickDefaultAccount(profilesController, defaultProvider).id
-        }
-
-      return CatalogFragmentFeed.create(
-        CatalogFeedArguments.CatalogFeedArgumentsLocalBooks(
-          filterAccount = filterAccountId,
-          ownership = CatalogFeedOwnership.CollectedFromAccounts,
-          searchTerms = null,
-          selection = FeedBooksSelection.BOOKS_FEED_LOANED,
-          sortBy = FeedFacet.FeedFacetPseudo.Sorting.SortBy.SORT_BY_TITLE,
-          title = context.getString(R.string.tabBooks)
-        )
-      )
-    }
-
-    private fun createCatalogFragment(
-      context: Context,
-      id: Int,
-      feedArguments: CatalogFeedArguments.CatalogFeedArgumentsRemote
-    ): Fragment {
-      logger.debug("[{}]: creating catalog fragment", id)
-      return CatalogFragmentFeed.create(feedArguments)
-    }
-  }
+  private val settingsConfiguration: BuildConfigurationServiceType
+): LifecycleObserver {
 
   init {
     this.navigationController.subscribeInfoStream(navigator)
@@ -308,54 +53,37 @@ class MainFragmentNavigationDelegate(
     LoggerFactory.getLogger(MainFragmentNavigationDelegate::class.java)
 
   private val subscriptions =
-    CompositeDisposable(
+    CompositeDisposable()
+
+  @OnLifecycleEvent(Lifecycle.Event.ON_START)
+  fun onStart() {
       this.navigationController.commandQueue
-        .subscribe(this::handleCommand),
-      this.navigationController.infoStream
-        .subscribe { action ->
-          this.logger.debug(action.toString())
-          this.onFragmentTransactionCompleted()
-        }
-    )
+        .subscribe(this::handleCommand)
+        .let { subscriptions.add(it) }
 
-  fun dispose() {
-    this.subscriptions.dispose()
+    this.navigationController.infoStream
+      .subscribe { action ->
+        this.logger.debug(action.toString())
+        this.onFragmentTransactionCompleted()
+      }
+      .let { subscriptions.add(it) }
   }
 
-  fun clearHistory() {
-    this.logger.debug("clearing bottom navigator history")
-    this.navigator.clearAll()
-  }
-
-  fun popBackStack(): Boolean {
-    return this.navigator.pop()
-  }
-
-  fun popToRoot(): Boolean {
-    val isAtRootOfStack = (1 == this.navigator.currentStackSize())
-    if (isAtRootOfStack) {
-      return false // Nothing to do
-    }
-    val currentTab = this.navigator.currentTab()
-    this.navigator.reset(currentTab, false)
-    return true
-  }
-
-  private fun backStackSize(): Int {
-    // Note: currentStackSize() is not safe to call here as it may throw an NPE.
-    return this.navigator.stackSize(this.navigator.currentTab()) - 1
+  @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+  fun onStop() {
+    subscriptions.clear()
   }
 
   private fun onFragmentTransactionCompleted() {
-    val isRoot = (0 == this.backStackSize())
+    val isRoot = (0 == this.navigator.backStackSize())
     this.logger.debug(
-      "controller stack size changed [{}, isRoot={}]", this.backStackSize(), isRoot
+      "controller stack size changed [{}, isRoot={}]", this.navigator.backStackSize(), isRoot
     )
     configureToolbar()
   }
 
   private fun configureToolbar() {
-    val isRoot = (0 == this.backStackSize())
+    val isRoot = (0 == this.navigator.backStackSize())
     this.activity.supportActionBar?.apply {
       setHomeAsUpIndicator(null)
       setHomeActionContentDescription(null)
@@ -363,12 +91,12 @@ class MainFragmentNavigationDelegate(
     }
   }
 
-  private fun handleCommand(command: TabbedNavigationCommand) {
+  fun handleCommand(command: TabbedNavigationCommand) {
     when(command) {
       TabbedNavigationCommand.AccountCommand.OnAccountCreated ->
-        popBackStack()
+        navigator.popBackStack()
       TabbedNavigationCommand.AccountCommand.OnSAMLEventAccessTokenObtained ->
-        popBackStack()
+        navigator.popBackStack()
       TabbedNavigationCommand.AccountCommand.OpenCatalogAfterAuthentication ->
         openCatalogAfterAuthentication()
       is TabbedNavigationCommand.AccountCommand.OpenErrorPage ->
@@ -381,7 +109,7 @@ class MainFragmentNavigationDelegate(
         openSettingsAccountRegistry()
 
       TabbedNavigationCommand.CatalogCommand.OnSAML20LoginSucceeded ->
-        popBackStack()
+        navigator.popBackStack()
       is TabbedNavigationCommand.CatalogCommand.OpenBookDetail ->
         openBookDetail(command.feedArguments, command.entry)
       is TabbedNavigationCommand.CatalogCommand.OpenBookDownloadLogin ->
