+++ simplified-main/src/main/java/org/nypl/simplified/main/MainFragmentViewModel.kt
@@ -10,7 +10,6 @@ import org.nypl.simplified.accounts.registry.api.AccountProviderRegistryType
 import org.nypl.simplified.buildconfig.api.BuildConfigurationServiceType
 import org.nypl.simplified.profiles.api.ProfileEvent
 import org.nypl.simplified.profiles.controller.api.ProfilesControllerType
-import org.nypl.simplified.ui.navigation.tabs.TabbedNavigationController
 
 /**
  * The view model for the main fragment.
